# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PlaisirXpress Adult Shop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  discountPrice?: number;
  category: string;
  categorySlug: string;
  imageUrl: string;
  images?: string[];
  colors?: string[];
  sizes?: string[];
  inStock: boolean;
  featured: boolean;
  rating: number;
  reviewCount: number;
  brand?: string;
  material?: string;
  length?: string;
  diameter?: string;
  tags?: string[];
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  imageUrl: string;
  productCount: number;
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
  color?: string;
  size?: string;
}

export interface Address {
  street: string;
  city: string;
  province: string;
  postalCode: string;
  country: string;
}

export interface CreateOrderInput {
  items: OrderItemInput[];
  customerEmail: string;
  customerName: string;
  shippingAddress: Address;
  paymentMethod: string;
  cardLastFour?: string;
}

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  price: number;
}

export interface Order {
  id: number;
  orderNumber: string;
  status: string;
  total: number;
  items: OrderItem[];
  createdAt: string;
}

export type ListProductsParams = {
  category?: string;
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  color?: string;
  featured?: boolean;
};