# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PlaisirXpress Adult Shop API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog
  - name: categories
    description: Product categories
  - name: cart
    description: Shopping cart
  - name: orders
    description: Orders
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: color
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        discountPrice:
          type: number
        category:
          type: string
        categorySlug:
          type: string
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        sizes:
          type: array
          items:
            type: string
        inStock:
          type: boolean
        featured:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        brand:
          type: string
        material:
          type: string
        length:
          type: string
        diameter:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - description
        - price
        - category
        - categorySlug
        - imageUrl
        - inStock
        - featured
        - rating
        - reviewCount

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        productCount:
          type: integer
      required:
        - id
        - name
        - slug
        - description
        - imageUrl
        - productCount

    CreateOrderInput:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
        customerEmail:
          type: string
        customerName:
          type: string
        shippingAddress:
          $ref: "#/components/schemas/Address"
        paymentMethod:
          type: string
        cardLastFour:
          type: string
      required:
        - items
        - customerEmail
        - customerName
        - shippingAddress
        - paymentMethod

    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        color:
          type: string
        size:
          type: string
      required:
        - productId
        - quantity

    Address:
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        province:
          type: string
        postalCode:
          type: string
        country:
          type: string
      required:
        - street
        - city
        - province
        - postalCode
        - country

    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
      required:
        - id
        - orderNumber
        - status
        - total
        - items
        - createdAt

    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - productId
        - productName
        - quantity
        - price