# lib/api-zod/src/generated/types/address.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PlaisirXpress Adult Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface Address {
  street: string;
  city: string;
  province: string;
  postalCode: string;
  country: string;
}