# lib/api-zod/src/generated/types/createOrderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PlaisirXpress Adult Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { Address } from "./address";
import type { OrderItemInput } from "./orderItemInput";

export interface CreateOrderInput {
  items: OrderItemInput[];
  customerEmail: string;
  customerName: string;
  shippingAddress: Address;
  paymentMethod: string;
  cardLastFour?: string;
}