# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PlaisirXpress Adult Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";

export interface Order {
  id: number;
  orderNumber: string;
  status: string;
  total: number;
  items: OrderItem[];
  createdAt: string;
}