# lib/api-zod/src/generated/types/orderItemInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PlaisirXpress Adult Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItemInput {
  productId: number;
  quantity: number;
  color?: string;
  size?: string;
}