# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PlaisirXpress Adult Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  discountPrice?: number;
  category: string;
  categorySlug: string;
  imageUrl: string;
  images?: string[];
  colors?: string[];
  sizes?: string[];
  inStock: boolean;
  featured: boolean;
  rating: number;
  reviewCount: number;
  brand?: string;
  material?: string;
  length?: string;
  diameter?: string;
  tags?: string[];
}